feat(perf): implement hpc optimizations for memory caching, lock-free inflight reads, block routing, and FUSE client
- Implement lock-free sharded MemoryCache using scc::HashIndex to prevent lock contention.
- Migrate InflightBlockReads to scc::HashIndex with drop-clean guards.
- Implement flat sorted Arc slice for dir_entry_cache with binary search lookups.
- Optimize FUSE readdir/readdirplus offset-prefiltering to skip entry building.
- Implement thread-local file handle caching for NVMe device.
- Align all metrics fields to distinct cache lines to prevent false sharing.
- Pin tokio worker threads to physical CPU cores.
- Add O_DIRECT fallback warning.
- Use zero-copy Bytes::from and Cow borrows on FUSE hot paths.
- Replace dynamic inode mutexes with static sharded StripeLocks.

+++ benches/high_concurrency_bench.rs
@@ -116,10 +116,7 @@ fn bench_squeezefs_concurrency(c: &mut Criterion) {
                     let router = router_ref.clone();
                     let data = data_ref.clone();
                     tokio::spawn(async move {
-                        router
-                            .write_file(&name, 0, &data, (base + i) as u64)
-                            .await
-                            .unwrap();
+                        router.write_file(&name, 0, &data, base + i).await.unwrap();
                         let mut con = router.dlm.get_connection().await.unwrap();
                         let _ = router.delete_file(&name, &mut con).await;
                     })
